Rearrange top-level components into "pages".
Following the Next.js style, take all of our top-level pages (as defined by
routes.jsx) and put them in their own directory. This will give us a better
sense of the application's entry points.
