+++ ui/__tests__/pages/requirements.test.js
@@ -1,21 +1,21 @@
 import { shallow } from 'enzyme';
 import React from 'react';
 
-import { PoliciesContainer } from '../../../components/policies/container';
+import { RequirementsContainer } from '../../pages/requirements';
 
-describe('<PoliciesContainer />', () => {
+describe('<RequirementsContainer />', () => {
   const location = { query: {
-    issuance__gt: '2015-01-01',
-    requirements__verb__icontains: 'must',
-    requirements__topics__id__in: '1,6,9',
+    policy__issuance__gt: '2015-01-01',
+    verb__icontains: 'must',
+    topics__id__in: '1,6,9',
     page: '5',
   } };
-  const pagedPolicies = {
+  const pagedReqs = {
     results: [{ thing: 1 }, { thing: 2 }],
     count: 2,
   };
   const result = shallow(React.createElement(
-    PoliciesContainer, { location, pagedPolicies }));
+    RequirementsContainer, { location, pagedReqs }));
 
   it('has a topics filter controls', () => {
     const controls = result.prop('filterControls');
@@ -23,10 +23,9 @@ describe('<PoliciesContainer />', () => {
     expect(controls[0].props.heading).toEqual('Topics');
 
     const selector = controls[0].props.selector;
-    expect(selector.props.insertParam).toEqual('requirements__topics__id__in');
+    expect(selector.props.insertParam).toEqual('topics__id__in');
     expect(selector.props.lookup).toEqual('topics');
   });
-
 
   describe('its tabs', () => {
     it('has two tabs with correct names', () => {
@@ -35,22 +34,22 @@ describe('<PoliciesContainer />', () => {
 
       const [reqTab, policyTab] = tabs;
 
-      expect(reqTab.props.active).toBeFalsy();
+      expect(reqTab.props.active).toBeTruthy();
       expect(reqTab.props.tabName).toBe('Requirements');
 
-      expect(policyTab.props.active).toBeTruthy();
+      expect(policyTab.props.active).toBeFalsy();
       expect(policyTab.props.tabName).toBe('Policies');
     });
 
-    it('transforms the query for the requirements tab', () => {
-      const reqTab = result.prop('tabs')[0];
-      const link = reqTab.props.link;
+    it('transforms the query for the policies tab', () => {
+      const policyTab = result.prop('tabs')[1];
+      const link = policyTab.props.link;
       expect(link).toEqual({
-        pathname: '/requirements',
+        pathname: '/policies',
         query: {
-          policy__issuance__gt: '2015-01-01',
-          verb__icontains: 'must',
-          topics__id__in: '1,6,9',
+          issuance__gt: '2015-01-01',
+          requirements__verb__icontains: 'must',
+          requirements__topics__id__in: '1,6,9',
         },
       });
     });
@@ -59,8 +58,8 @@ describe('<PoliciesContainer />', () => {
   it('has correct pageContent', () => {
     const pageContent = result.prop('pageContent');
 
-    expect(pageContent.props.policies).toEqual(pagedPolicies.results);
-    expect(pageContent.props.count).toEqual(pagedPolicies.count);
+    expect(pageContent.props.requirements).toEqual(pagedReqs.results);
+    expect(pageContent.props.count).toEqual(pagedReqs.count);
   });
 
   it('has configured selectedFilters', () => {
@@ -72,3 +71,4 @@ describe('<PoliciesContainer />', () => {
     expect(selectedFilters.props.query).toEqual(location.query);
   });
 });
+
