+++ ui/pages/privacy.jsx
@@ -98,3 +98,4 @@ export default function PrivacyView() {
     </div>
   );
 }
+
